requirement.txt
pandas
matplotlib.pyplot
numpy
streamlit
serpapi